requirements.txt
pandas
yfinance
python-dateutil
pydrive
tqdm
pymysql
python-dotenv